Refactor modules and move editor utilities
