+++ Plugins/Character2D/Source/Character2DRuntime/Public/Character2DRuntimeModule.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "Modules/ModuleManager.h"
 
-class FCharacter2DRuntimeModule : public IModuleInterface
+class CHARACTER2DRUNTIME_API FCharacter2DRuntimeModule : public IModuleInterface
 {
 public:
     virtual void StartupModule() override {}
